requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==19.3.0
boto3==1.16.57
botocore==1.19.57
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
coverage==4.5.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0
django-braces==1.14.0
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.10.2
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==4.3.0
Faker==0.9.1
hashids==1.3.1
idna==2.10
importlib-metadata==1.6.1
jmespath==0.10.0
Markdown==3.2.2
mixer==6.1.3
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
text-unidecode==1.2
twilio==6.51.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
